# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Management System API
  version: 1.0.0
  description: |
    API documentation for the Medical Management System backend.

    This API provides endpoints for managing users, doctors, patients, appointments, and medical records.
    Authentication is required for most endpoints using JWT tokens.
  contact:
    name: Medical Management System
    email: support@medical-system.com
  license:
    name: MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.medical-system.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users (admin only)
      operationId: getAllUsers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: role
          in: query
          schema:
            type: string
            enum:
              - admin
              - doctor
              - patient
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: Forbidden - admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID (admin only)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Users
      summary: Update user (admin only)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - Users
      summary: Delete user (admin only)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /users/{id}/deactivate:
    put:
      tags:
        - Users
      summary: Deactivate user (admin only)
      operationId: deactivateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /users/{id}/activate:
    put:
      tags:
        - Users
      summary: Activate user (admin only)
      operationId: activateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  # Doctor endpoints
...
  /medical-records:
    post:
      tags:
        - Medical Records
      summary: Create a new medical record (doctor only)
      operationId: createMedicalRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMedicalRecord'
      responses:
        '201':
          description: Medical record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecord'
    get:
      tags:
        - Medical Records
      summary: Get medical records (filtered by user role)
      operationId: getMedicalRecords
      parameters:
        - name: patient_id
          in: query
          schema:
            type: integer
        - name: doctor_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicalRecord'
  /medical-records/{id}:
    get:
      tags:
        - Medical Records
      summary: Get medical record by ID
      operationId: getMedicalRecordById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Medical record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecord'
        '404':
          description: Medical record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Medical Records
      summary: Update medical record (doctor only)
      operationId: updateMedicalRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMedicalRecord'
      responses:
        '200':
          description: Medical record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalRecord'
    delete:
      tags:
        - Medical Records
      summary: Delete medical record (doctor/admin only)
      operationId: deleteMedicalRecord
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Medical record deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /medical-records/patient/{patient_id}:
    get:
      tags:
        - Medical Records
      summary: Get patient's medical history
      operationId: getPatientMedicalHistory
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient's medical history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicalRecord'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfileUpdate:
      type: object
      required:
        - first_name
        - last_name
      properties:
        first_name:
          type: string
          minLength: 2
          maxLength: 50
          example: John
        last_name:
          type: string
          minLength: 2
          maxLength: 50
          example: Doe
        phone:
          type: string
          example: '+1234567890'
        date_of_birth:
          type: string
          format: date
          example: '1990-01-01'
        gender:
          type: string
          enum:
            - male
            - female
            - other
          example: male
        address:
          type: string
          example: 123 Main St, City, State 12345
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum:
            - admin
            - doctor
            - patient
          example: patient
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        phone:
          type: string
          example: '+1234567890'
        date_of_birth:
          type: string
          format: date
          example: '1990-01-01'
        gender:
          type: string
          enum:
            - male
            - female
            - other
          example: male
        address:
          type: string
          example: 123 Main St, City, State 12345
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        license_number:
          type: string
          example: MD123456
        specialization:
          type: string
          example: Cardiology
        experience_years:
          type: integer
          example: 10
        consultation_fee:
          type: number
          format: decimal
          example: 150
        education:
          type: string
          example: MD from Harvard Medical School
        bio:
          type: string
          example: Experienced cardiologist with 10 years of practice
        is_available:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        patient_id:
          type: integer
          example: 1
        doctor_id:
          type: integer
          example: 1
        appointment_date:
          type: string
          format: date
          example: '2023-12-25'
        appointment_time:
          type: string
          format: time
          example: '10:00'
        duration:
          type: integer
          example: 30
        status:
          type: string
          enum:
            - scheduled
            - confirmed
            - cancelled
            - completed
            - no_show
          example: scheduled
        appointment_type:
          type: string
          enum:
            - consultation
            - follow_up
            - emergency
            - routine_checkup
          example: consultation
        reason:
          type: string
          example: Regular checkup
        notes:
          type: string
          example: Patient has been feeling chest pain
        reminder_sent:
          type: boolean
          example: false
        cancelled_by:
          type: integer
          example: 1
        cancelled_at:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        cancellation_reason:
          type: string
          example: Patient unavailable
        patient:
          $ref: '#/components/schemas/User'
        doctor:
          $ref: '#/components/schemas/Doctor'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
...
    MedicalRecord:
      type: object
      properties:
        id:
          type: integer
          example: 1
        patient_id:
          type: integer
          example: 1
        doctor_id:
          type: integer
          example: 1
        appointment_id:
          type: integer
          example: 1
        visit_date:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        chief_complaint:
          type: string
          example: Chest pain
        symptoms:
          type: string
          example: Sharp pain in chest, shortness of breath
        diagnosis:
          type: string
          example: Mild chest strain
        treatment_plan:
          type: string
          example: Rest, pain medication, follow-up in 1 week
        medications_prescribed:
          type: string
          example: Ibuprofen 400mg twice daily
        follow_up_instructions:
          type: string
          example: Return if symptoms worsen
        vital_signs:
          type: object
          properties:
            blood_pressure:
              type: string
              example: 120/80
            heart_rate:
              type: integer
              example: 72
            temperature:
              type: number
              example: 98.6
            respiratory_rate:
              type: integer
              example: 16
            oxygen_saturation:
              type: integer
              example: 98
        lab_results:
          type: string
          example: Blood work normal
        notes:
          type: string
          example: Patient responded well to treatment
        next_appointment_recommended:
          type: boolean
          example: true
        patient:
          $ref: '#/components/schemas/User'
        doctor:
          $ref: '#/components/schemas/Doctor'
        appointment:
          $ref: '#/components/schemas/Appointment'
        createdAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
    CreateMedicalRecord:
      type: object
      required:
        - patient_id
      properties:
        patient_id:
          type: integer
          example: 1
        appointment_id:
          type: integer
          example: 1
        visit_date:
          type: string
          format: date-time
          example: '2023-01-01T10:00:00Z'
        chief_complaint:
          type: string
          example: Chest pain
        symptoms:
          type: string
          example: Sharp pain in chest, shortness of breath
        diagnosis:
          type: string
          example: Mild chest strain
        treatment_plan:
          type: string
          example: Rest, pain medication, follow-up in 1 week
        medications_prescribed:
          type: string
          example: Ibuprofen 400mg twice daily
        follow_up_instructions:
          type: string
          example: Return if symptoms worsen
        vital_signs:
          type: object
          properties:
            blood_pressure:
              type: string
              example: 120/80
            heart_rate:
              type: integer
              example: 72
            temperature:
              type: number
              example: 98.6
            respiratory_rate:
              type: integer
              example: 16
            oxygen_saturation:
              type: integer
              example: 98
        lab_results:
          type: string
          example: Blood work normal
        notes:
          type: string
          example: Patient responded well to treatment
        next_appointment_recommended:
          type: boolean
          example: true

    # Common schemas
    Success:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
        success:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid input provided
        message:
          type: string
          example: The request could not be processed
        success:
          type: boolean
          example: false
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 10
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false